Pass Go test flags and environment variables (#13)
Pass Go test flags and environment variables

To more closely duplicate the test environment that GopherJS has
created for use with NodeJS, accept Go test flags on the command line
and pass them, as well as the environment variables, as part of the
javascript that is sent to the browser.

Among other things, this enables the use of the -v, -run and -bench
flags. And it provides the code being tested access to the environment
variables through the normal os package mechanism.

If having the environment variables copied into the javascript that gets
loaded into the browser is not desirable, this can be disabled with the
new -env boolean flag. (-env=0 disables environment variables being
copied.)